# packages/api-client/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fullstack Starter API",
    "description": "API documentation for Fullstack Starter Backend",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/notes/": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "title": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "content": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "content",
                        "createdAt"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "content": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "content",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/notes/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "null"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "title": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "content": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "content",
                        "createdAt",
                        "updatedAt"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "title": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "content": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "content",
                        "updatedAt"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}